+++ setup.cfg
@@ -4,8 +4,8 @@
 # Big example here: https://github.com/pallets/flask/blob/main/setup.cfg
 
 [metadata]
-name = ode_lab
-version = attr: ode_lab.__version__
+name = ode-lab-base
+version = attr: ode_lab.base.__version__
 
 [options]
 packages = find_namespace:
@@ -19,10 +19,11 @@ install_requires =
     numpy
     xarray
     tqdm
+    ipykernel
 
 
 [options.packages.find]
-include = ode_lab.*
+include = ode-lab.base.*
 
 [options.extras_require]
 dev =
